--- packages/cli/src/UserManagement/controllers/index.ts
@@ -1,6 +1,0 @@
-/* eslint-disable import/no-cycle */
-export * from './AuthController';
-export * from './MeController';
-export * from './OwnerController';
-export * from './PasswordResetController';
-export * from './UserController';
